scripts/requirements.txt
openai>=1.75.0
azure-identity>=1.19.0
azure-ai-projects>=1.0.0b11
requests>=2.32.0
python-dotenv>=1.0.1
rich>=13.9.0
